return_paths_ops=False and skip calcs if files exist

+++ project/ana_cov_comp/python/get_1pt_ewin_alms.py
@@ -49,31 +49,34 @@ canonized_combos = {}
 # S
 # iterate over all canonized windows
 for field_info1 in field_infos:
-    ewin_name1, ewin_paths1, ewin_ops1 = psc.get_ewin_info_from_field_info(field_info1, d, mode='w')
+    ewin_name1, ewin_paths1, ewin_ops1 = psc.get_ewin_info_from_field_info(
+        field_info1, d, mode='w', return_paths_ops=True
+        )
     
     alm_fn = f'{ewin_alms_dir}/{ewin_name1}_alm.npy'
 
     if ewin_name1 not in canonized_combos:
         canonized_combos[ewin_name1] = [field_info1]
-        assert not os.path.isfile(alm_fn), \
-            f'{alm_fn} exists but we should not yet have produced it in loop'
-        log.info(ewin_name1)
+        if os.path.isfile(alm_fn):
+            log.info(f'{alm_fn} exists, skipping')
+        else:
+            log.info(ewin_name1)
 
-        ewin1_data = 1
-        for i, path1 in enumerate(ewin_paths1):
-            ewin1 = so_map.read_map(path1)
-            ewin1.data = psc.optags2ops[ewin_ops1[i]](ewin1.data)
-            ewin1_data *= ewin1.data
-        ewin1.data = ewin1_data
-        ewin1_data = None
+            ewin1_data = 1
+            for i, path1 in enumerate(ewin_paths1):
+                ewin1 = so_map.read_map(path1)
+                ewin1.data = psc.optags2ops[ewin_ops1[i]](ewin1.data)
+                ewin1_data *= ewin1.data
+            ewin1.data = ewin1_data
+            ewin1_data = None
 
-        # note we are going to calculate window alm to 2lmax, but that's ok because
-        # the lmax_limit is half the Nyquist limit
-        assert lmax <= ewin1.get_lmax_limit(), \
-            "the requested lmax is too high with respect to the map pixellisation"
+            # note we are going to calculate window alm to 2lmax, but that's ok because
+            # the lmax_limit is half the Nyquist limit
+            assert lmax <= ewin1.get_lmax_limit(), \
+                "the requested lmax is too high with respect to the map pixellisation"
 
-        alm = sph_tools.map2alm(ewin1, niter=niter, lmax=2*lmax)
-        np.save(alm_fn, alm)
+            alm = sph_tools.map2alm(ewin1, niter=niter, lmax=2*lmax)
+            np.save(alm_fn, alm)
     else:
         canonized_combos[ewin_name1].append(field_info1)
         assert os.path.isfile(alm_fn), \
@@ -83,34 +86,37 @@ for field_info1 in field_infos:
 # N
 # iterate over all canonized windows
 for field_info1 in field_infos:
-    ewin_name1, ewin_paths1, ewin_ops1 = psc.get_ewin_info_from_field_info(field_info1, d, mode='ws', extra='sqrt_pixar')
+    ewin_name1, ewin_paths1, ewin_ops1 = psc.get_ewin_info_from_field_info(
+        field_info1, d, mode='ws', extra='sqrt_pixar', return_paths_ops=True
+        )
     
     alm_fn = f'{ewin_alms_dir}/{ewin_name1}_alm.npy'
 
     if ewin_name1 not in canonized_combos:
         canonized_combos[ewin_name1] = [field_info1]
-        assert not os.path.isfile(alm_fn), \
-            f'{alm_fn} exists but we should not yet have produced it in loop'
-        log.info(ewin_name1)
+        if os.path.isfile(alm_fn):
+            log.info(f'{alm_fn} exists, skipping')
+        else:
+            log.info(ewin_name1)
 
-        ewin1_data = 1
-        for i, path1 in enumerate(ewin_paths1):
-            ewin1 = so_map.read_map(path1)
-            ewin1.data = psc.optags2ops[ewin_ops1[i]](ewin1.data)
-            ewin1_data *= ewin1.data
-        ewin1.data = ewin1_data
-        ewin1_data = None
+            ewin1_data = 1
+            for i, path1 in enumerate(ewin_paths1):
+                ewin1 = so_map.read_map(path1)
+                ewin1.data = psc.optags2ops[ewin_ops1[i]](ewin1.data)
+                ewin1_data *= ewin1.data
+            ewin1.data = ewin1_data
+            ewin1_data = None
 
-        # we know we need to multiply by sqrt_pixar
-        ewin1.data *= ewin1.data.pixsizemap()**0.5
+            # we know we need to multiply by sqrt_pixar
+            ewin1.data *= ewin1.data.pixsizemap()**0.5
 
-        # note we are going to calculate window alm to 2lmax, but that's ok because
-        # the lmax_limit is half the Nyquist limit
-        assert lmax <= ewin1.get_lmax_limit(), \
-            "the requested lmax is too high with respect to the map pixellisation"
+            # note we are going to calculate window alm to 2lmax, but that's ok because
+            # the lmax_limit is half the Nyquist limit
+            assert lmax <= ewin1.get_lmax_limit(), \
+                "the requested lmax is too high with respect to the map pixellisation"
 
-        alm = sph_tools.map2alm(ewin1, niter=niter, lmax=2*lmax)
-        np.save(alm_fn, alm)
+            alm = sph_tools.map2alm(ewin1, niter=niter, lmax=2*lmax)
+            np.save(alm_fn, alm)
     else:
         canonized_combos[ewin_name1].append(field_info1)
         assert os.path.isfile(alm_fn), \
